+++ test/my_relative.c
@@ -7,16 +7,16 @@
   conditions of the GPLv2 as it is applied to this software, see the
   FLOSS License Exception
   <http://www.mysql.com/about/legal/licensing/foss-exception.html>.
-  
+
   This program is free software; you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published
   by the Free Software Foundation; version 2 of the License.
-  
+
   This program is distributed in the hope that it will be useful, but
   WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
   or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
   for more details.
-  
+
   You should have received a copy of the GNU General Public License along
   with this program; if not, write to the Free Software Foundation, Inc.,
   51 Franklin St, Fifth Floor, Boston, MA 02110-1301  USA
@@ -700,7 +700,7 @@ DECLARE_TEST(t_rows_fetched_ptr1)
     rc = SQLFetchScroll(hstmt,SQL_FETCH_NEXT,0);
     while (rc == SQL_SUCCESS || rc == SQL_SUCCESS_WITH_INFO)
     {
-      fprintf(stdout,"total rows fetched: %ld\n", rowsFetched);
+      fprintf(stdout,"total rows fetched: %ld\n", (long)rowsFetched);
       myassert(rowsFetched == rowsSize);
       i++; rowsFetched= 0;
       rc = SQLFetchScroll(hstmt, SQL_FETCH_NEXT, 0);
@@ -722,7 +722,7 @@ DECLARE_TEST(t_rows_fetched_ptr1)
     rc = SQLFetchScroll(hstmt,SQL_FETCH_NEXT,0);
     while (rc == SQL_SUCCESS || rc == SQL_SUCCESS_WITH_INFO)
     {
-      fprintf(stdout,"total rows fetched: %ld\n", rowsFetched);
+      fprintf(stdout,"total rows fetched: %ld\n", (long)rowsFetched);
       myassert(rowsFetched == rowsSize);
       i++;rowsFetched= 0;
       rc = SQLFetchScroll(hstmt,SQL_FETCH_NEXT,0);
